fix(ai-config): drop per-task in-memory cache, always read from DB
Stale-cache failure mode in multi-task deployments: the per-task
Map cache was invalidated only on the task that handled the
update request. Other tasks served the old config for up to 5
minutes per TTL. After a BYOK provider switch, ~1-1/N of requests
went to the previous provider's endpoint with the previous key.

Drop the cache. The query is a single indexed findFirst on
organizationId returning one small row from the same VPC; the
cost is dominated by the network round-trip, well under a
millisecond. If profiling later shows the lookup matters,
shared cache + cross-task invalidation is the right answer
(local map plus per-task TTL is not).

Renames the module from ai-config-cache to ai-config-loader to
match its current responsibility.
